tests/wpt/tests/dom/nodes/Document-contentType/contentType/contenttype_gif.html

<!DOCTYPE html>
<title>GIF document.contentType === 'image/gif'</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
  var iframe = document.createElement('iframe');
  iframe.addEventListener('load', this.step_func_done(function() {
    assert_equals(iframe.contentDocument.contentType, "image/gif");
  }), false);
  iframe.src = "../resources/t.gif";
  document.body.appendChild(iframe);
});
</script>
